Add title and aria-label
