+++ libs/components/src/dialog/dialog.module.ts
@@ -2,6 +2,8 @@ import { DialogModule as CdkDialogModule } from "@angular/cdk/dialog";
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 
+import { I18nPipe } from "../shared";
+
 import { DialogService } from "./dialog.service";
 import { DialogComponent } from "./dialog/dialog.component";
 import { DialogCloseDirective } from "./directives/dialog-close.directive";
@@ -15,6 +17,7 @@ import { SimpleDialogComponent } from "./simple-dialog/simple-dialog.component";
     DialogComponent,
     DialogTitleContainerDirective,
     SimpleDialogComponent,
+    I18nPipe,
   ],
   exports: [CdkDialogModule, DialogComponent, SimpleDialogComponent],
   providers: [DialogService],
