+++ libs/components/src/dialog/dialog.service.stories.ts
@@ -2,7 +2,11 @@ import { DialogModule, DialogRef, DIALOG_DATA } from "@angular/cdk/dialog";
 import { Component, Inject } from "@angular/core";
 import { Meta, moduleMetadata, Story } from "@storybook/angular";
 
+import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
+
 import { ButtonModule } from "../button";
+import { I18nPipe } from "../shared";
+import { I18nMockService } from "../utils/i18n-mock.service";
 
 import { DialogService } from "./dialog.service";
 import { DialogComponent } from "./dialog/dialog.component";
@@ -64,9 +68,20 @@ export default {
         DialogComponent,
         DialogTitleContainerDirective,
         StoryDialogContentComponent,
+        I18nPipe,
       ],
       imports: [ButtonModule, DialogModule],
-      providers: [DialogService],
+      providers: [
+        DialogService,
+        {
+          provide: I18nService,
+          useFactory: () => {
+            return new I18nMockService({
+              close: "Close",
+            });
+          },
+        },
+      ],
     }),
   ],
   parameters: {
