+++ libs/components/src/dialog/dialog/dialog.component.html
@@ -8,7 +8,12 @@
     <h1 bitDialogTitleContainer class="tw-mb-0 tw-grow tw-text-lg tw-uppercase">
       <ng-content select="[bitDialogTitle]"></ng-content>
     </h1>
-    <button class="tw-border-0 tw-bg-transparent tw-p-0" bitDialogClose>
+    <button
+      bitDialogClose
+      class="tw-border-0 tw-bg-transparent tw-p-0"
+      title="{{ 'close' | i18n }}"
+      attr.aria-label="{{ 'close' | i18n }}"
+    >
       <i class="bwi bwi-close tw-text-xs tw-font-bold tw-text-main" aria-hidden="true"></i>
     </button>
   </div>
