+++ libs/components/src/dialog/dialog/dialog.stories.ts
@@ -1,6 +1,10 @@
 import { Meta, moduleMetadata, Story } from "@storybook/angular";
 
+import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
+
 import { ButtonModule } from "../../button";
+import { I18nPipe } from "../../shared";
+import { I18nMockService } from "../../utils/i18n-mock.service";
 import { DialogCloseDirective } from "../directives/dialog-close.directive";
 import { DialogTitleContainerDirective } from "../directives/dialog-title-container.directive";
 
@@ -12,7 +16,17 @@ export default {
   decorators: [
     moduleMetadata({
       imports: [ButtonModule],
-      declarations: [DialogTitleContainerDirective, DialogCloseDirective],
+      declarations: [DialogTitleContainerDirective, DialogCloseDirective, I18nPipe],
+      providers: [
+        {
+          provide: I18nService,
+          useFactory: () => {
+            return new I18nMockService({
+              close: "Close",
+            });
+          },
+        },
+      ],
     }),
   ],
   args: {
