+++ libs/components/src/dialog/directives/dialog-close.directive.ts
@@ -1,18 +1,15 @@
 import { DialogRef } from "@angular/cdk/dialog";
-import { Directive, Input, Optional } from "@angular/core";
+import { Directive, HostBinding, Input, Optional } from "@angular/core";
 
 @Directive({
   selector: "[bitDialogClose]",
-  host: {
-    "(click)": "close()",
-  },
 })
 export class DialogCloseDirective {
   @Input("bit-dialog-close") dialogResult: any;
 
   constructor(@Optional() public dialogRef: DialogRef<any>) {}
 
-  close() {
+  @HostBinding() close() {
     this.dialogRef.close(this.dialogResult);
   }
 }
